requirements.txt
google-api-python-client==2.127.0
python-dotenv==1.0.1
pandas==2.2.2
SQLAlchemy==1.4.49
psycopg2-binary==2.9.9
apache-airflow==2.9.0
apache-airflow-providers-postgres==5.10.0
loguru==0.7.2

